+++ hoover/contrib/__init__.py
@@ -1,0 +1,4 @@
+from django.conf import settings
+
+class installed:
+    twofactor = ('hoover.contrib.twofactor' in settings.INSTALLED_APPS)