+++ CHANGES.md
@@ -2,6 +2,7 @@
 * upgrade to Django 1.10
 * "External" loader that fetches documents from a URL
 * set no-cache and no-referral response headers
+* TOTP login - `hoover.contrib.twofactor`
 
 ## 0.1 (2016-04-28)
 * initial release